login + refactor
Добавил авторизацию, поправил ошибки с перемещением формы, провел рефакторинг кода.
