+++ WindowsFormsApp1/CrimeaCloud.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.33328.57
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsFormsApp1", "WindowsFormsApp1.csproj", "{F7152B2A-F5C6-4719-B356-F06D1A507D39}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CrimeaCloud", "CrimeaCloud.csproj", "{F7152B2A-F5C6-4719-B356-F06D1A507D39}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
