+++ WindowsFormsApp1/RegisterForm.cs
@@ -1,20 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-using System.Net.Http;
-using System.Net.Http.Json;
-using System.Text.Json;
-using System.Threading;
 
 
-
-namespace WindowsFormsApp1
+namespace CrimeaCloud
 {
     public partial class RegisterForm : Form
     {
@@ -24,27 +13,6 @@ namespace WindowsFormsApp1
             InitializeComponent();
 
         }
-        private void LoginField_TextChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        private void PasswordField_TextChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        private void email_TextChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        private void PssswordCon_TextChanged(object sender, EventArgs e)
-        {
-
-        }
-
-
         private void RegisterForm_MouseMove(object sender, MouseEventArgs e)
         {
             if (e.Button == MouseButtons.Left)
@@ -66,13 +34,6 @@ namespace WindowsFormsApp1
 
         }
 
-        private void button2_Click(object sender, EventArgs e)
-        {
-            Form1 f1 = new Form1();
-            f1.Show();
-            Hide();
-        }
-
         private void RegisterForm_Load(object sender, EventArgs e)
         {
 
@@ -83,7 +44,9 @@ namespace WindowsFormsApp1
 
         private void AutorReg_Click(object sender, EventArgs e)
         {
-
+            Hide();
+            LoginForm loginform = new LoginForm();
+            loginform.Show();
         }
 
         private async void RegReg_Click(object sender, EventArgs e)
@@ -96,46 +59,12 @@ namespace WindowsFormsApp1
                 password = bunifuTextBox2.Text,
                 confirmPassword = bunifuTextBox3.Text
             };
-            var response = await RegisterUserAsync(data);
-            //MessageBox.Show(response);
+            var response = await ConnectHttp.RegisterUserAsync(data);
+            Console.WriteLine(response.Content.ReadAsStringAsync().Result);
         }
 
-        public static async Task<string> RegisterUserAsync(object data)
-        {
-            var httpClient = new HttpClient();
-            httpClient.BaseAddress = new Uri("http://176.99.11.107/api/user/");
-            var json = JsonSerializer.Serialize(data);
-            Console.WriteLine(json);
-            var content = new StringContent(json, Encoding.UTF8, "application/json");
-            try
-            {
-                var response = httpClient.PostAsync("signup", content).Result;
-                //Thread.Sleep(10000);
-                Console.WriteLine(response.Content.ReadAsStringAsync().Result);
-                //response.EnsureSuccessStatusCode();
-                //return await response.Content.ReadAsStringAsync();
-            }
-            catch (HttpRequestException ex)
-            {
-                Console.WriteLine($"HTTP error occurred: {ex.Message}");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex);
-            }
+        
 
-            return null;
-        }
-
-        private void pictureBox3_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void LoginRegist_TextChanged(object sender, EventArgs e)
-        {
-
-        }
     }
 
 }
