+++ WindowsFormsApp1/RegisterForm.resx
@@ -118,6 +118,43 @@
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
   <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
+  <data name="pictureBox5.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
+    <value>
+        iVBORw0KGgoAAAANSUhEUgAAAGAAAABgCAYAAADimHc4AAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8
+        YQUAAAAJcEhZcwAACxMAAAsTAQCanBgAAAdDSURBVHhe7Z1JqBxVFIYzqFFjAi40JBpFNEFJYkQlgmA2
+        KogDgsaNiLhyExAFx0U2TlEMQSUqrgQDGhxwQmLAuBAcFyqJgqAEx4gKgolGcYr/qfpeL/L61HCqu191
+        1/3gh+6q//731u336lRX1as3K5FIJBKJRCKRSCQSiUQikUgkEolE4lAOHjy4SnpE+lT6Teo6Ngc2Fw9L
+        K5mmwaPwedLj0r9Soj82N49JRzBtg0GBNvlvSYlq7JQG9yEo7IksNlGHLUxfMxRk+/y026nPP9IKpjGO
+        QqzgJmJsZhrjKOSzPCsRYDfTGEch+/OsRIB9TGMcglywdRamwQVbHHJcsHUWpsEFWxxyXLB1FqbBBVsc
+        clywdRamwQVbHHKKeE06CXtn0DYvlp62CSgCexxyyvhdukOaS7OJRds4R7pR+lUqhWZxyKnKx9Iamk4c
+        2rbV0vu2oVWhaRxy6mCnLZ6UFhAx9mhbjpYekP6WakFEHHIifC+tI2Zs0TZcLn1lGxSBmDjkNGEsi7TG
+        XKnIlkFcHHKaMjZFWmOsVWTLIDYOOYOi1UVaY6tdZMsgOg45RVwhfZ2/rETrirTGMlVk7Rx+VfZK1+cv
+        fegiDjkueCIb0IoirTHULbJTP0ALaV9I1kkTyHHBlqG3Z0kfZCuqMyNFWn1Giuwn0nlEZOSLfbDFIccF
+        Ww8tmipi+2x9RUZWpNVHpMhOje8wYnpkawvAFoccF2zT0Kol0vOZqTpDLdLKjv6GnkzENHKLD7Y45Lhg
+        c5Flxou0ssJFlgiX3OqDLQ45LtgKkS1apK8hIowy7AcgXGTLyFoUgC0OOS7YKiH7yIq02tgusHGRLSNv
+        5oMtDjku2CqjJkMt0vI0KbK1DwKy1gVgi0OOC7baqOnAi7TWDbzIlpFH+GCLQ44LtjCKiBbp3j5ar4dW
+        ZMvIo3ywxSHHBVsjFLNAstu760zgN9KVyF5XxfqwvgZylJUlFoAtDjku2AaC4iK7kDoM/HtGHuuDLQ45
+        LtgGhiIjRbqMoX3TztILwBaHHBdsA0fRVqSfyzppRqMiW0behQ+2OOS4YBsa6qJukZ7Cvsg1LrJl5F35
+        YItDjgu2oaJu6hzl1Pom25SsxwKwxSHHBdtIUHdlRXrkV9zybn2wxSHHBdvIUJf9vumO7HT2oWS9F4At
+        Djku2EaOul4qvYyWsnjkqO9CsMUhxwVbZ2EaXLDFIccFW2dhGlywxSHHBVtnYRpcsMUhxwVbZ2EaXLDF
+        IccFW2dhGlywxSHHBVtnYRpcsMUhxwVbZ2EaXLDFIccFW2dhGlywxSHHBVur0TBnS3YL4jOSndj7U7LT
+        3XYRfpN0OtbaqG0h2OKQ44KttWiIy6UPs8H62Ak8eyLMkTSrTNa6AGxxyHHB1ko0vAukn7KBVuM9aRHN
+        K5E388EWhxwXbK1DQztb+iMbZD12SfOJKSVv4oMtDjku2FqFhmUX+b/MBhhjK1Gl4HfBFoccF2ytQsPa
+        kI+uERcSVwheF2xxyHHB1ho0pPnSL9ngmvEukYXgdcEWhxwXbK1BQ1qXj2wgnEmsCz4XbHGUMcjbQwbB
+        RQytL1r/VG5zsaOiU6QbpP9sQQH3Eesij/3GLZMuk+yxnnYzQA9scZTRtkeWncrQ+qL1H+U2l7uxmnd7
+        vshlJ9bKqI094nO99IMFsDiOMuw2vjZxDEPri9aX3RW9Gqt5r8oXuezFWhu1PV7awds4Clkp1blnc9jM
+        Y2h90fq/cptL7wPU6+PyRS4HsIZQ+zm8bIaC7HG8baHwibRaX/aQwd4uTK/th6uIRh/AwNBAjpDscbxt
+        YDHD6ovWl9WsDVjNuzVf5LIH68yjwdiHsEWa6d3RuQypL1r/em5zsaMgO3I5Xyo7CnqV2PagQa2QNku7
+        pZl4puh1DKUvWn9XbivkdsnuoivjJmITVdGk2e2LZVR58JJ5pp0Z1TI7z3SLZLdI2jfun6Vt0jIs3UYT
+        YRdf9khN2U5kDy1bK3l3atuuLX0Ihibi1mxKmrGWuAy9v1SyK2lFbMPebTQRx0pN/r3KS0Rl6P0aqUpe
+        391WJ9FE3JtNSX3sIs5yYiznBCk7rVCRh2jabTQRC6Ufsympx3oiLOMoyS5T1uGAtISI7qJJuFj61mak
+        Bq9Is2l/uPSiLQzwhjTyv1VoBdpwuwviWZuFAPaPik6UzpHetAUNeEHq3COd7QMYJ/Yz7MmBDRsXdjHs
+        yYENGwV2I1fZqe8yNjHsyYENGzb30NfN+dsQ9p3htGzQk0S+bYXYbYhlZ0E9rN39dJWh9/Y3yRE2EjFZ
+        sHEe70h2hGOnFercsmjY8f21dNNDy+zc00apzodqj0+YzMPUfPum8bZkj7XJjvENvV4k2U9vlTOjdvH+
+        DJr2Resvkb4wcwH2IT0q1b4JeGzQxtk5m8+lHdJtUuG/E9R6u0XlTsk+pO8kK6z2DdqeIf2gtAprKfLO
+        la6W7Eqb3Rppf0Busmsn9h8JezcE5Mya9T/h49qjiA3hAgAAAABJRU5ErkJggg==
+</value>
+  </data>
   <data name="pictureBox4.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
         iVBORw0KGgoAAAANSUhEUgAAADIAAAAyCAYAAAAeP4ixAAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAL
